Move search-related files to Search subfolder.
