+++ ILSpy/Analyzers/Builtin/TypeExtensionMethodsAnalyzer.cs
