+++ ILSpy/Search/SearchPane.cs
