+++ ILSpy/Search/SearchPane.xaml
